BENCH: Rework benchmarks
